Refactor: Split mainSets.js into individual workout type files
The mainSets.js file was becoming too large and difficult to manage.
This change splits the file into smaller, more focused files, each
containing the logic for a single workout type (e.g., ENDURANCE_BASE,
MAX_SPRINT).

- Created a new directory `lib/data/mainSets/`.
- Moved each workout type definition from `mainSetDefinitions` in
  `lib/data/mainSets.js` into its own file within the new directory
  (e.g., `lib/data/mainSets/ENDURANCE_BASE.js`).
- Copied necessary constants (e.g., `en1Distances`, `sp2Distances`)
  into their respective new files.
- Updated `lib/data/mainSets.js` to import these individual workout
  type functions and re-export them in the `mainSetDefinitions` object.

This refactoring improves code organization and maintainability without
changing the overall functionality.
